+++ Resource.h
@@ -1,7 +1,6 @@
 #pragma once
-#include "badDirectX.h"
-#include "EasyDirectX.h"
-#include "EasyDirectXUtils.h"
+
+#include <d3d12.h>
 #include <wrl/client.h>
 
 class CommittedResource
@@ -13,23 +12,23 @@ public:
     Microsoft::WRL::ComPtr<ID3D12Resource> load(
         ID3D12Device4* device,
         ID3D12GraphicsCommandList2* cl,
-        const void* data, SIZE_T element_count, SIZE_T type_size);
+        const void* data, UINT64 element_count, UINT64 type_size);
 
 
     ID3D12Resource* get() const noexcept;
 
-    constexpr SIZE_T count() const noexcept
+    constexpr UINT64 count() const noexcept
     {
         return mCount;
     }
 
-    constexpr SIZE_T stride() const noexcept
+    constexpr UINT64 stride() const noexcept
     {
         return mStride;
     }
 
 private:
     Microsoft::WRL::ComPtr<ID3D12Resource> mResource;
-    SIZE_T                                 mCount = 0;
-    SIZE_T                                 mStride = 0;
+    UINT64                            mCount = 0;
+    UINT64                            mStride = 0;
 };
